Factored out middlewares, added custom request type
